Hide our shame in a separate file
