+++ src/Update/Update.csproj
@@ -78,6 +78,7 @@
       <Link>Properties\SolutionAssemblyInfo.cs</Link>
     </Compile>
     <Compile Include="AnimatedGifWindow.cs" />
+    <Compile Include="AuthenticodeTools.cs" />
     <Compile Include="CopStache.cs" />
     <Compile Include="Mono.Options\Options.cs" />
     <Compile Include="Program.cs" />
